Divide exception tests into 2 collections
