+++ test/Telegram.Bot.Tests.Integ/Framework/TestCollectionOrderer.cs
@@ -18,8 +18,6 @@ namespace Telegram.Bot.Tests.Integ.Framework
             Constants.TestCollections.DeleteMessage,
             Constants.TestCollections.Games,
             Constants.TestCollections.Payment,
-
-            // ToDo Tests that can be fully automated or divided into 2 collections:
             Constants.TestCollections.ChatMemberAdministration,
             Constants.TestCollections.Exceptions,
 
@@ -46,6 +44,7 @@ namespace Telegram.Bot.Tests.Integ.Framework
             Constants.TestCollections.Games2,
             Constants.TestCollections.SupergroupAdminBots,
             Constants.TestCollections.ChannelAdminBots,
+            Constants.TestCollections.Exceptions2,
         };
 
         public IEnumerable<ITestCollection> OrderTestCollections(IEnumerable<ITestCollection> testCollections)
